requirements.txt
opencv-python==4.8.1.78
pytesseract==0.3.10
python-dateutil==2.8.2
icalendar==5.0.7
Pillow==10.0.1
tkinter-tooltip==2.0.0
customtkinter==5.2.0 